controls are component not resource
